Configuration Manager added The configuration manager uses ini files and enables configuration to be inheritable for two layers for now : the app layer and the Module layer. The container manages automatically injecting the config to the model, the controller and the view in business/conf/apps.ini you can set parameters using the sections [controller] to set configs to the controller [module] to set configs to the module [view] to set configs to the view [cache] to set configs to the cache
you can redefine the settings in the business/<yourmodule>/conf/<yourmodule>.ini
Configuration is automatically accessible in the controller, view and module by using $this->configuration. this will
allow you to access the relative sections.
The controller Can access all the configurations options using $this->Config which represents the ConfigurationManager
object.
The ConfigurationManager Object is accessible like an array (it implements ArrayAccess) so you can access for example :
$this->Config['controller']['yourconfig1']

The level support is now only to 3 levels :
[controller]
yourconfig1 = avalue
yourconfig2[Yoursubconfig3] = anothervalue
yourconfig2[Yoursubconfig4] = athirdvalue
you can access the configuration params using :
$this->Config['controller']['yourconfig2']['Yoursubconfig3']

2. The ConfigurationParser :
The configurationParser allows you to parse and use any Ini file in your application. It loads an INI file and parse it.
Like the configurationManager, it is also accessible as an array. This class can be handy for components etc.

The configuration is cached at the container layer.